GoalPrefab, PlayerCamera, Menu Script,
Changed:
Player Camera values
PlayerCharacter mass and angular drag
Re-added SceneSwitch.script to GolfGoalPost.Prefab
UI was reduced to information on how to active the pause menu to declutter the screen.
Unchecked scenes that did not need to be built in version 1 of electronic prototype.

Removed:
Text from HowToPlay.Scene that might have confused players as the features mentioned were not yet implemented